Refactor confetti code to module import
