+++ index.html
@@ -42,7 +42,7 @@
         </div>
     </div>
     <!-- Script -->
-    <script src="script.js"></script>
+    <script src="script.js" type="module"></script>
 </body>
 
 </html>
